Added hooks for enheter, virksomhetsdata and sykefraværsstatistikk
