+++ src/utils/enhetsregisteretContext.tsx
@@ -1,19 +1,6 @@
-import React, { createContext, FunctionComponent, useEffect, useState } from 'react';
-import { RestRessurs, RestStatus } from '../api/api-utils';
-import {
-    hentInformasjonOmOverordnetEnhet,
-    hentInformasjonOmUnderenhet,
-    OverordnetEnhet,
-    RestOverordnetEnhet,
-    RestUnderenhet,
-    Underenhet,
-} from '../api/enhetsregisteret-api';
-import { useOrgnr } from '../hooks/useOrgnr';
-
-export interface EnhetsregisteretState {
-    restUnderenhet: RestUnderenhet;
-    restOverordnetEnhet: RestOverordnetEnhet;
-}
+import React, { createContext, FunctionComponent } from 'react';
+import { RestStatus } from '../api/api-utils';
+import { EnhetsregisteretState, useEnheter } from '../hooks/useEnheter';
 
 export const enhetsregisteretContext = createContext<EnhetsregisteretState>({
     restUnderenhet: {
@@ -24,73 +11,14 @@ export const enhetsregisteretContext = createContext<EnhetsregisteretState>({
     },
 });
 
-interface DataForVirksomhet<T> {
-    orgnr: string;
-    restData: RestRessurs<T>;
-}
-
-export const useRestDataForFlereVirksomheter = <T extends Object>(
-    hentData: (orgnr: string) => Promise<RestRessurs<T>>,
-    orgnr: string | undefined
-): [RestRessurs<T>, DataForVirksomhet<T>[]] => {
-    const [gjeldendeData, setGjeldendeData] = useState<RestRessurs<T>>({
-        status: RestStatus.IkkeLastet,
-    });
-
-    const [dataForAlleVirksomheter, setDataForAlleVirksomheter] = useState<DataForVirksomhet<T>[]>(
-        []
-    );
-
-    useEffect(() => {
-        if (orgnr) {
-            const dataOmGjeldendeVirksomhet = dataForAlleVirksomheter.find(
-                (data) => data.orgnr === orgnr
-            );
-
-            if (dataOmGjeldendeVirksomhet) {
-                setGjeldendeData(dataOmGjeldendeVirksomhet.restData);
-            } else {
-                setGjeldendeData({
-                    status: RestStatus.IkkeLastet,
-                });
-                const hentRestDataOgSetState = async () => {
-                    const restData = await hentData(orgnr);
-                    setGjeldendeData(restData);
-                    setDataForAlleVirksomheter([
-                        ...dataForAlleVirksomheter,
-                        {
-                            orgnr: orgnr,
-                            restData: restData,
-                        },
-                    ]);
-                };
-                hentRestDataOgSetState();
-            }
-        }
-    }, [orgnr, dataForAlleVirksomheter, hentData]);
-
-    return [gjeldendeData, dataForAlleVirksomheter];
-};
-
 export const EnhetsregisteretProvider: FunctionComponent = (props) => {
-    const underenhetOrgnr = useOrgnr();
-
-    const [gjeldendeUnderenhet] = useRestDataForFlereVirksomheter<Underenhet>(
-        hentInformasjonOmUnderenhet,
-        underenhetOrgnr
-    );
-    const [gjeldendeOverordnetEnhet] = useRestDataForFlereVirksomheter<OverordnetEnhet>(
-        hentInformasjonOmOverordnetEnhet,
-        gjeldendeUnderenhet.status === RestStatus.Suksess
-            ? gjeldendeUnderenhet.data.overordnetEnhet
-            : undefined
-    );
+    const { restOverordnetEnhet, restUnderenhet } = useEnheter();
 
     const Provider = enhetsregisteretContext.Provider;
 
     const contextValue: EnhetsregisteretState = {
-        restUnderenhet: gjeldendeUnderenhet,
-        restOverordnetEnhet: gjeldendeOverordnetEnhet,
+        restUnderenhet,
+        restOverordnetEnhet,
     };
 
     return <Provider value={contextValue}>{props.children}</Provider>;
